+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-19T05:02:55.012Z",
+    "2026-05-19T07:57:33.807Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {